chore: move syncTransaction to sync file
