+++ packages/element-lib/src/func/func.js
@@ -3,7 +3,6 @@ const base64url = require('base64url');
 const crypto = require('crypto');
 const secp256k1 = require('secp256k1');
 const multihashes = require('multihashes');
-const { isTransactionValid, isBatchFileValid, isAnchorFileValid } = require('../sidetree/utils/validation');
 
 // This function applies f, an async function, sequentially to an array of values
 // We need it because:
@@ -73,66 +72,6 @@ const readThenWriteToCache = async (sidetree, hash) => {
   return record;
 };
 
-const syncTransaction = async (sidetree, transaction, onlyDidUniqueSuffix = null) => {
-  if (process.env.NODE_ENV !== 'test') {
-    console.info('sync', transaction.transactionNumber);
-  }
-  try {
-    isTransactionValid(transaction);
-    const anchorFile = await readThenWriteToCache(sidetree, transaction.anchorFileHash);
-    isAnchorFileValid(anchorFile);
-    // Only sync the batch files containing operations about that didUniqueSuffix if provided
-    if (onlyDidUniqueSuffix && !anchorFile.didUniqueSuffixes.includes(onlyDidUniqueSuffix)) {
-      return null;
-    }
-    const batchFile = await sidetree.storage.read(anchorFile.batchFileHash);
-    isBatchFileValid(batchFile);
-    const operations = batchFileToOperations(batchFile);
-    const [transactionWithTimestamp] = await sidetree.blockchain
-      .extendSidetreeTransactionWithTimestamp([transaction]);
-    const operationsByDidUniqueSuffixes = operations.map((operation) => {
-      const didUniqueSuffix = getDidUniqueSuffix(operation.decodedOperation);
-      return {
-        type: didUniqueSuffix,
-        didUniqueSuffix,
-        transaction: transactionWithTimestamp,
-        operation,
-      };
-    });
-    const filteredOperationByDidUniqueSuffixes = operationsByDidUniqueSuffixes
-      // Only keep operations related to the didUniqueSuffix if provided
-      .filter(op => !onlyDidUniqueSuffix || op.didUniqueSuffix === onlyDidUniqueSuffix);
-    const writeOperationToCache = (op) => {
-      const operationId = `operation:${op.operation.operationHash}${op.transaction.transactionTime}`;
-      return sidetree.db.write(operationId, op);
-    };
-    return executeSequentially(
-      writeOperationToCache,
-      filteredOperationByDidUniqueSuffixes,
-    ).then(() => {
-      if (operationsByDidUniqueSuffixes.length !== filteredOperationByDidUniqueSuffixes.length) {
-        return null;
-      }
-      return sidetree.db.write(`transaction:${transaction.transactionNumber}`, {
-        type: 'transaction',
-        ...transaction,
-      });
-    });
-  } catch (error) {
-    console.log(error);
-    // https://stackoverflow.com/questions/18391212/is-it-not-possible-to-stringify-an-error-using-json-stringify
-    const stringifiedError = JSON.stringify(
-      error,
-      Object.getOwnPropertyNames(error),
-    );
-    return sidetree.db.write(`transaction:${transaction.transactionNumber}`, {
-      type: 'transaction',
-      ...transaction,
-      error: stringifiedError,
-    });
-  }
-};
-
 // TODO check is signatures are the same as sidetree's
 const signEncodedPayload = (encodedHeader, encodedPayload, privateKey) => {
   const toBeSigned = `${encodedHeader}.${encodedPayload}`;
@@ -179,7 +118,6 @@ module.exports = {
   payloadToHash,
   getDidUniqueSuffix,
   batchFileToOperations,
-  syncTransaction,
   signEncodedPayload,
   verifyOperationSignature,
   readThenWriteToCache,
